requirements.txt

ffmpeg-python